added lifecycleObject to reduce boilerplate. failed to come up with an elegant expandable solution on creature damage model with dynamic limb structure so ended up with hardcoded limbs for humanoids for now

+++ Assets/Scripts/Selskiyvrach/VampireHunter/Gameplay/Model/Bullets/IBullet.cs
@@ -2,7 +2,7 @@
 {
     public interface IBullet
     {
-        int Damage { get; }
+        float Damage { get; }
         void Launch(BulletLaunchData launchData);
     }
 }
